+++ src/inventory/inventory.h
@@ -9,17 +9,20 @@
 
 
 #define inventory_selected(inv) \
-    inv->items->get(inv->items, inv->selected)
+    inv->items[inv->selected]
 
 #define inventory_grid_size(inventory) ceil(sqrt(inventory->max_size))
 
 #define inventory_shortcut_offset(input) (input % KEY_INVENTORY_SHORTCUT_FIRST)
 
+#define inventory_at_shortcut(inventory, input) inventory->items[(input % KEY_INVENTORY_SHORTCUT_FIRST)]
+
 
 typedef struct Inventory {
     uint16_t max_size;
     uint16_t selected;
-    List *items;
+    uint16_t count;
+    Item **items;
     Grid *grid;
 } Inventory;
 
@@ -40,9 +43,11 @@ void inventory_use_selected(Inventory *inventory, Player *player);
 
 void inventory_grid_update(Inventory *inventory);
 
-void inventory_close(PanelInputEvent *event);
+void inventory_panel_close(PanelInputEvent *event);
 
-void inventory_navigate(PanelInputEvent *event);
+void inventory_panel_navigate(PanelInputEvent *event);
+
+void inventory_panel_drop_selected(PanelInputEvent *event);
 
 void inventory_display(Inventory *inventory);