+++ src/util/event.c
@@ -31,52 +31,24 @@ static Listener LISTENERS[EVENT_TYPE_NUM][MAX_EVENT_LISTENER_PER_EVENT] = {
         (Listener) item_light_source_place
     },
     [EVENT_PANEL_CLOSE] = {
-        (Listener) inventory_close
+        (Listener) inventory_panel_close
     }
 };
 
 static InputListener INPUT_LISTENERS[MAX_INPUT_LISTENER] = {
-    {
-        .executor = (Listener) player_move,
-        .inputs = {KEY_NORTH, KEY_EAST, KEY_SOUTH, KEY_WEST}
-    },
-    {
-        .executor = (Listener) mouse_handler,
-        .inputs = {KEY_MOUSE}
-    },
-    {
-        .executor = (Listener) inventory_player_use_selected,
-        .inputs = {KEY_USE_SELECTED}
-    },
-    {
-        .executor = (Listener) inventory_player_shortcut_select,
-        .inputs = {KEY_INVENTORY_SHORTCUTS}
-    },
-    {
-        .executor = (Listener) inventory_player_display,
-        .inputs = {KEY_PLAYER_INVENTORY}
-    },
-    {
-        .executor = (Listener) ncurses_resize,
-        .inputs = {KEY_RESIZE}
-    }
+    {(Listener) player_move,                      {KEY_NORTH, KEY_EAST, KEY_SOUTH, KEY_WEST}},
+    {(Listener) mouse_handler,                    {KEY_MOUSE}},
+    {(Listener) inventory_player_use_selected,    {KEY_USE_SELECTED}},
+    {(Listener) inventory_player_shortcut_select, {KEY_INVENTORY_SHORTCUTS}},
+    {(Listener) inventory_player_display,         {KEY_PLAYER_INVENTORY}},
+    {(Listener) inventory_player_drop_shortcut,   {KEY_DROP_ITEM}},
+    {(Listener) ncurses_resize,                   {KEY_RESIZE}}
 };
 
 static PanelInputListener PANEL_INPUT_LISTENERS[MAX_INPUT_LISTENER] = {
-    {
-        .base = {
-            .executor = (Listener) inventory_navigate,
-            .inputs = {KEY_NORTH, KEY_EAST, KEY_SOUTH, KEY_WEST, KEY_USE_SELECTED}
-        },
-        .types = {PANEL_INVENTORY}
-    },
-    {
-        .base = {
-            .executor = (Listener) inventory_player_set_shortcut,
-            .inputs = {KEY_INVENTORY_SHORTCUTS}
-        },
-        .types = {PANEL_INVENTORY}
-    }
+    {{(Listener) inventory_panel_navigate,      {KEY_NORTH, KEY_EAST, KEY_SOUTH, KEY_WEST, KEY_USE_SELECTED}}, {PANEL_INVENTORY}},
+    {{(Listener) inventory_player_set_shortcut, {KEY_INVENTORY_SHORTCUTS}},                                    {PANEL_INVENTORY}},
+    {{(Listener) inventory_panel_drop_selected, {KEY_DROP_ITEM}},                                              {PANEL_INVENTORY}}
 };