+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ###### NEW:
 * [x] Tools will wear out after using them
-* [ ] Items can be dropped from the inventory
+* [x] Items can be dropped from the inventory
 * [x] Items can be placed on shortcut from the inventory
 * [x] i386 tar build
 
@@ -15,5 +15,6 @@ and reduce the lookup time significantly.
 ###### FIXED:
 * [x] Clear out some really nasty stack smashing bugs
 * [x] When the inventory opened, the selected item is not marked
+* [ ] Items are added to the first empty slot
 * [x] Tar build did not copy the env variables
 * [x] Bunch of Valgrind errors